+++ tests/deprecated_api/test_remove_1-6.py
@@ -19,7 +19,6 @@ import torch
 from torch.optim import Optimizer
 
 from pytorch_lightning import Trainer
-from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.callbacks.early_stopping import EarlyStopping
 from pytorch_lightning.plugins import PrecisionPlugin
 from pytorch_lightning.plugins.training_type import DDPPlugin
@@ -165,11 +164,6 @@ def test_v1_6_0_deprecated_disable_validation():
     trainer = Trainer()
     with pytest.deprecated_call(match="disable_validation` is deprecated in v1.4"):
         _ = trainer.disable_validation
-
-
-def test_v1_6_0_every_n_val_epochs():
-    with pytest.deprecated_call(match="use `every_n_epochs` instead"):
-        _ = ModelCheckpoint(every_n_val_epochs=1)
 
 
 def test_v1_6_0_deprecated_hpc_load(tmpdir):